uint8 changed to uint32

+++ contracts/Lottery.sol
@@ -7,8 +7,8 @@ contract Lottery {
   mapping(address => uint32[]) soldTickets; // Map holding tickets and buyer addresses
   mapping(uint32 => address payable[]) playersByTicket; // Map holding tickets and buyer addresses
   uint256 public amountPerAddress;
-  uint8 public winningNr;
-  uint8[] allPlayedNumbers;
+  uint32 public winningNr;
+  uint32[] allPlayedNumbers;
   address payable[] allPlayers;
   address payable[] winners;
 
@@ -17,7 +17,7 @@ contract Lottery {
   }
 
   function buyTicket(uint32 _ticketNr) public payable {
-    require(_ticketNr < 251 && _ticketNr > 0, "Ticket number must be in range [0, 250]"); // Prevent overflow of uint8
+    require(_ticketNr < 251 && _ticketNr > 0, "Ticket number must be in range [0, 250]");
     require(msg.value >= 1, "A ticket costs 1 Ether"); // Checks if enough ether is sent to buy a lottery ticket
     soldTickets[msg.sender].push(_ticketNr);
     playersByTicket[_ticketNr].push(msg.sender);
